Move cell validation to updateTableStyle function.
Instead of a separate expression, allow apps to change the styling f a cell based on CSS class, and use a designated function for this purpose.

On the client side, setData() invalidates any td styling, so we can't just set it and forget. Instead, we have to store the classes to be applied to each cell in a 'cellClasses' object and re-apply the classes every time the data or the classes change. This will likely need to be optimized pretty soon.