Updated tests to expect numeric to be converted to text.
We don't really have a good way to express/enforce a numeric data type between HTable and R (integer only? Positive/negative? Rounding? etc.)

It would take a lot more thought to get that right. Also, how do we deal with class conversions of data.frames or matrices? Certainly not obvious, so for now, just treat it all as text.